- added flag to allow/deny new moves, the flag is disabled when there's an arrow key press and reset when the game status is updated.
This improves the animation of the tiles to finish the sliding smoothly without overlapping (i.e. ignoring quick key-presses).

NOTE: I couldn't find a way to detect when ALL the slide animation finish (i.e. the LAST animation finishes), so the reset of the flag is performed when the delayed timer completes to add a new tile to the game.